LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/addEventListener-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: undefined_variable
CONSOLE MESSAGE: SyntaxError: Unexpected token ')'

PASS window.onerror - addEventListener
